building dir rename
