--- packages/fe-tw/src/components/Pages/BuildingDetailPage/BuildingVotings.tsx
@@ -1,16 +1,0 @@
-import { BuildingVoteItem } from "./BuildingVoteItem";
-
-export const BuildingVotings = (props: { votings: number[] }) => {
-    return (
-        <div className="flex flex-col mt-10">
-            <article className="prose">
-                <h2>Voting Items</h2>
-            </article>
-            <div className="flex flex-col">
-                {props.votings?.map(vote => (
-                    <BuildingVoteItem key={vote} voteId={vote} />
-                ))}
-            </div>
-        </div>
-    )
-};
