+++ packages/fe-tw/src/components/Pages/BuildingDetailsPage/BuildingVoteItem.tsx
@@ -1,15 +1,16 @@
 import { ReusableAvatar } from "@/components/Avatars/ReusableAvatar";
-import { activeProposals } from "@/consts/props";
+import { activeProposals } from "@/consts/proposals";
 import { ClockIcon } from "@/resources/icons/ClockIcon";
-import { TimeLabel } from "@/components/Typography/TimeLabel";
+import moment from "moment";
 
 export const BuildingVoteItem = ({ voteId }: { voteId: number }) => {
-    const vote = activeProposals.find(proposal => proposal.id === voteId);
+    const vote = activeProposals.find(proposal => proposal.id === voteId)
 
     return (
         <div className="flex flex-row mt-5">
             <ReusableAvatar
                 imageAlt={vote?.title!}
+                imageSource={vote?.imageUrl}
                 size="md"
                 isRounded
                 isFocusAvailable={false}
@@ -21,9 +22,11 @@ export const BuildingVoteItem = ({ voteId }: { voteId: number }) => {
                 </article>
                 <div className="flex flex-row items-center">
                     <ClockIcon />
-                    <TimeLabel date={vote?.started as Date} formatType="dateAsTimeRange" />
+                    <span className="text-xs ml-2 text-slate-700">
+                        {moment(vote?.started).format('dddd, LT')}
+                    </span>
                 </div>
             </div>
         </div>
     );
-};
+}
