+++ packages/fe-tw/src/components/Wallets/WalletConnectModal.tsx
@@ -4,7 +4,6 @@ import { useWalletInterface } from "@/services/useWalletInterface";
 import { shortEvmAddress } from "@/services/util";
 import { connectToMetamask } from "@/services/wallets/metamask/MetaMaskClient";
 import { openWalletConnectModal } from "@/services/wallets/walletconnect/WalletConnectClient";
-import { Button } from "react-daisyui";
 
 export function WalletConnectModal() {
   const { accountId, walletInterface } = useWalletInterface();
@@ -32,7 +31,7 @@ export function WalletConnectModal() {
         <>
           <div className="line-clamp-1 mx-4">{shortEvmAddress(accountId)}</div>
           <button
-            className="bg-purple-700 text-white px-4 py-2 rounded hover:bg-purple-900 transition"
+            className="bg-purple-700 text-white px-4 py-2 rounded-full hover:bg-purple-900 transition"
             onClick={() => {
               walletInterface?.disconnect();
               metaMaskCtx.setMetamaskAccountAddress("");
@@ -44,7 +43,7 @@ export function WalletConnectModal() {
       ) : (
         <>
           <button
-            className="bg-purple-700 text-white px-4 py-2 rounded hover:bg-purple-900 transition"
+            className="bg-purple-700 text-white px-4 py-2 rounded-full hover:bg-purple-900 transition"
             onClick={handleOpenModal}
           >
             Connect Wallet
@@ -62,7 +61,7 @@ export function WalletConnectModal() {
                 <h2 className="text-lg font-bold">Connect Wallet</h2>
                 <div className="flex flex-col gap-4 mt-4">
                   <button
-                    className="bg-purple-700 text-white px-4 py-2 rounded hover:bg-purple-900 transition w-60"
+                    className="bg-purple-700 text-white px-4 py-2 rounded-full hover:bg-purple-900 transition w-60"
                     onClick={() => {
                       connectToMetamask();
                       handleCloseModal();
@@ -71,7 +70,7 @@ export function WalletConnectModal() {
                     MetaMask
                   </button>
                   <button
-                    className="bg-purple-700 text-white px-4 py-2 rounded hover:bg-purple-900 transition w-60"
+                    className="bg-purple-700 text-white px-4 py-2 rounded-full hover:bg-purple-900 transition w-60"
                     onClick={() => {
                       openWalletConnectModal();
                       handleCloseModal();
@@ -82,7 +81,7 @@ export function WalletConnectModal() {
                 </div>
                 <div className="flex justify-end mt-4">
                   <button
-                    className="bg-gray-200 px-4 py-2 rounded hover:bg-gray-300 transition"
+                    className="bg-gray-200 px-4 py-2 rounded-full hover:bg-gray-300 transition"
                     onClick={handleCloseModal}
                   >
                     Close
